requirements.txt
streamlit
pdfplumber
pymupdf
python-docx
pytesseract
Pillow
keybert
transformers
torch
sentence-transformers